+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/api_inference/mining.py
@@ -1,20 +1,18 @@
 import json
 import os
-import time
 
 import re
 from bfcl_eval.model_handler.model_style import ModelStyle
 from bfcl_eval.model_handler.utils import (
-    func_doc_language_specific_pre_processing,
-    retry_with_backoff,
+    func_doc_language_specific_pre_processing
 )
-from openai import OpenAI, RateLimitError
-from bfcl_eval.model_handler.api_inference.openai import OpenAIHandler
+from openai import OpenAI
+from bfcl_eval.model_handler.api_inference.openai_completion import OpenAICompletionsHandler
 
-class MiningHandler(OpenAIHandler):
+class MiningHandler(OpenAICompletionsHandler):
     def __init__(self, model_name, temperature) -> None:
         super().__init__(model_name, temperature)
-        self.model_style = ModelStyle.OpenAI
+        self.model_style = ModelStyle.OpenAI_Completions
         self.client = OpenAI(
             base_url= os.getenv("MINING_BASE_URL"),
             api_key=os.getenv("MINING_API_KEY"),