+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/api_inference/novita.py
@@ -1,11 +1,11 @@
 import os
 
-from bfcl_eval.model_handler.api_inference.openai import OpenAIHandler
+from bfcl_eval.model_handler.api_inference.openai_completion import OpenAICompletionsHandler
 from bfcl_eval.model_handler.model_style import ModelStyle
 from openai import OpenAI
 
 
-class NovitaHandler(OpenAIHandler):
+class NovitaHandler(OpenAICompletionsHandler):
     def __init__(self, model_name, temperature) -> None:
         super().__init__(model_name, temperature)
         self.model_style = ModelStyle.NOVITA_AI