+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/api_inference/openai_completion.py
@@ -1,9 +1,10 @@
 import json
 import os
 import time
+from typing import Any
 
-from bfcl_eval.model_handler.base_handler import BaseHandler
 from bfcl_eval.constants.type_mappings import GORILLA_TO_OPENAPI
+from bfcl_eval.model_handler.base_handler import BaseHandler
 from bfcl_eval.model_handler.model_style import ModelStyle
 from bfcl_eval.model_handler.utils import (
     convert_to_function_call,
@@ -18,10 +19,10 @@ from bfcl_eval.model_handler.utils import (
 from openai import OpenAI, RateLimitError
 
 
-class OpenAIHandler(BaseHandler):
+class OpenAICompletionsHandler(BaseHandler):
     def __init__(self, model_name, temperature) -> None:
         super().__init__(model_name, temperature)
-        self.model_style = ModelStyle.OpenAI
+        self.model_style = ModelStyle.OpenAI_Completions
         self.client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
 
     def decode_ast(self, result, language="Python"):
@@ -56,34 +57,17 @@ class OpenAIHandler(BaseHandler):
         tools = inference_data["tools"]
         inference_data["inference_input_log"] = {"message": repr(message), "tools": tools}
 
+        kwargs = {
+            "messages": message,
+            "model": self.model_name.replace("-FC", ""),
+            "temperature": self.temperature,
+            "store": False,
+        }
+
         if len(tools) > 0:
-            # Reasoning models don't support temperature parameter
-            # Beta limitation: https://platform.openai.com/docs/guides/reasoning/beta-limitations
-            if "o1" in self.model_name or "o3-mini" in self.model_name:
-                return self.generate_with_backoff(
-                    messages=message,
-                    model=self.model_name.replace("-FC", ""),
-                    tools=tools,
-                )
-            else:
-                return self.generate_with_backoff(
-                    messages=message,
-                    model=self.model_name.replace("-FC", ""),
-                    temperature=self.temperature,
-                    tools=tools,
-                )
-        else:
-            if "o1" in self.model_name or "o3-mini" in self.model_name:
-                return self.generate_with_backoff(
-                    messages=message,
-                    model=self.model_name.replace("-FC", ""),
-                )
-            else:
-                return self.generate_with_backoff(
-                    messages=message,
-                    model=self.model_name.replace("-FC", ""),
-                    temperature=self.temperature,
-                )
+            kwargs["tools"] = tools
+
+        return self.generate_with_backoff(**kwargs)
 
     def _pre_query_processing_FC(self, inference_data: dict, test_entry: dict) -> dict:
         inference_data["message"] = []
@@ -163,7 +147,7 @@ class OpenAIHandler(BaseHandler):
         return inference_data
 
     def _add_reasoning_content_if_available_FC(
-        self, api_response: any, response_data: dict
+        self, api_response: Any, response_data: dict
     ) -> None:
         """
         OpenAI models don't show reasoning content in the api response,
@@ -212,19 +196,12 @@ class OpenAIHandler(BaseHandler):
     def _query_prompting(self, inference_data: dict):
         inference_data["inference_input_log"] = {"message": repr(inference_data["message"])}
 
-        # OpenAI reasoning models don't support temperature parameter
-        # Beta limitation: https://platform.openai.com/docs/guides/reasoning/beta-limitations
-        if "o1" in self.model_name or "o3-mini" in self.model_name:
-            return self.generate_with_backoff(
-                messages=inference_data["message"],
-                model=self.model_name,
-            )
-        else:
-            return self.generate_with_backoff(
-                messages=inference_data["message"],
-                model=self.model_name,
-                temperature=self.temperature,
-            )
+        return self.generate_with_backoff(
+            messages=inference_data["message"],
+            model=self.model_name,
+            temperature=self.temperature,
+            store=False,
+        )
 
     def _pre_query_processing_prompting(self, test_entry: dict) -> dict:
         functions: list = test_entry["function"]
@@ -279,7 +256,7 @@ class OpenAIHandler(BaseHandler):
         return inference_data
 
     def _add_reasoning_content_if_available_prompting(
-        self, api_response: any, response_data: dict
+        self, api_response: Any, response_data: dict
     ) -> None:
         """
         OpenAI models don't show reasoning content in the api response,