+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/utils.py
@@ -70,7 +70,8 @@ def convert_to_tool(functions, mapping, model_style):
     oai_tool = []
     for item in functions:
         if "." in item["name"] and model_style in [
-            ModelStyle.OpenAI,
+            ModelStyle.OpenAI_Completions,
+            ModelStyle.OpenAI_Responses,
             ModelStyle.Mistral,
             ModelStyle.Google,
             ModelStyle.OSSMODEL,
@@ -165,8 +166,15 @@ def convert_to_tool(functions, mapping, model_style):
         ]:
             oai_tool.append(item)
         elif model_style in [
+            ModelStyle.OpenAI_Responses
+        ]:
+            oai_tool.append({"type": "function", 
+                             "name": item["name"], 
+                             "description": item["description"], 
+                             "parameters": item["parameters"]})
+        elif model_style in [
             ModelStyle.COHERE,
-            ModelStyle.OpenAI,
+            ModelStyle.OpenAI_Completions,
             ModelStyle.Mistral,
             ModelStyle.FIREWORK_AI,
             ModelStyle.WRITER,